Trying float instead of double

+++ cpp_headers/rbt_classes.hpp
@@ -86,8 +86,8 @@ public:
 class TerminalHistory : public History
 {
 public:
-    std::vector<double> reward;
-    TerminalHistory(std::vector<int>& history, std::vector<double> reward = {0.0, 0.0});
+    std::vector<float> reward;
+    TerminalHistory(std::vector<int>& history, std::vector<float> reward = {0.0, 0.0});
     TerminalHistory copy();
     void set_reward();
 };
@@ -103,14 +103,14 @@ class Policy
 {
 public:
     char player;
-    std::unordered_map<std::string, std::vector<double> > policy_dict;
+    std::unordered_map<std::string, std::vector<float> > policy_dict;
     Policy();
     Policy(char player, std::string& file_path);
-    Policy(char player, std::unordered_map<std::string, std::vector<double> >& policy_dict);
+    Policy(char player, std::unordered_map<std::string, std::vector<float> >& policy_dict);
     Policy copy();
     void load_policy(char player, std::string& file_path);
-    void update_policy_for_given_information_set(InformationSet& information_set, std::vector<double>& prob_distribution);
-    std::unordered_map<std::string, std::vector<double> > read_policy_from_json(std::string& file_path, char player);
+    void update_policy_for_given_information_set(InformationSet& information_set, std::vector<float>& prob_distribution);
+    std::unordered_map<std::string, std::vector<float> > read_policy_from_json(std::string& file_path, char player);
 };
 
 // similar to python's split method
